+++ components/calendar/index.tsx
@@ -1,0 +1,313 @@
+"use client";
+import React, { useState, useRef, useEffect } from "react";
+import FullCalendar from "@fullcalendar/react";
+import dayGridPlugin from "@fullcalendar/daygrid";
+import interactionPlugin from "@fullcalendar/interaction";
+import moment from "moment";
+import { cn } from "@/lib/utils";
+import { checkSlotAvailability } from "@/utils/api-calls";
+import "./styles.css";
+import { Button } from "@/components/button";
+import { useRouter } from "next/navigation";
+
+const CalendarPage = () => {
+  const router = useRouter();
+  const [events, setEvents] = useState<any[]>([]);
+  const [originalSlots, setOriginalSlots] = useState(null);
+  const [filter, setFilter] = useState("AM");
+  const lastFetchedDate = useRef("");
+  const startMonth = new Date();
+  const [isButtonDisabled, setIsButtonDisabled] = useState<boolean>(true);
+  const [selectedDateTime, setSelectedDateTime] = useState<any>(null);
+  const [hasSelectedTime, setHasSelectedTime] = useState(false);
+
+  // NOTE: Commenting it for now
+
+  // console.log("slot", originalSlots);
+  // const formatEvents = (slotsData: any, filter: string) => {
+  //   console.log("slotsData", slotsData);
+  //   const filteredEvents = Object.values(slotsData)
+  //     .filter((slot: any) => slot?.reservations && slot?.slots)
+  //     .flatMap((slot: any) =>
+  //       slot?.slots
+  //         .filter((timeSlot: any) => {
+  //           const hour = parseInt(timeSlot.time.split(":")[0], 10);
+  //           return filter === "AM" ? hour < 12 : hour >= 12;
+  //         })
+  //         .filter((timeSlot: any) => !timeSlot?.available)
+  //         .map((timeSlot: any) => {
+  //           const combinedDateTime = moment(slot.date).set({
+  //             hour: parseInt(timeSlot.time.split(":")[0], 10),
+  //             minute: parseInt(timeSlot.time.split(":")[1], 10),
+  //             second: 0,
+  //             millisecond: 0,
+  //           });
+
+  //           return {
+  //             title: `Booked - ${combinedDateTime.format("h:mm A")}`,
+  //             start: combinedDateTime.toISOString(),
+  //             allDay: true,
+  //             color: "green",
+  //           };
+  //         }),
+  //     );
+
+  //   setEvents(filteredEvents);
+  // };
+
+  const formatEvents = (slotsData: any, filter: string) => {
+    const filteredEvents = Object.values(slotsData)
+      .filter((slot: any) => slot?.reservations && slot?.slots)
+      .flatMap((slot: any) =>
+        slot?.slots
+          .filter((timeSlot: any) => {
+            const hour = parseInt(timeSlot.time.split(":")[0], 10);
+            return filter === "AM" ? hour < 12 : hour >= 12;
+          })
+          .filter((timeSlot: any) => !timeSlot?.available)
+          .map((timeSlot: any) => {
+            console.log("slot.date", slot.date);
+            const baseDate = new Date(slot.date);
+            const [hours, minutes] = timeSlot.time.split(":").map(Number);
+
+            const combinedDateTime = new Date(baseDate);
+            combinedDateTime.setUTCHours(hours, minutes, 0, 0);
+
+            console.log(
+              "Debug - Combined DateTime:",
+              combinedDateTime.toISOString(),
+            );
+
+            const utcTime = new Date(combinedDateTime).toLocaleTimeString(
+              "en-US",
+              {
+                timeZone: "UTC",
+                hour: "numeric",
+                minute: "2-digit",
+                hour12: true,
+              },
+            );
+
+            console.log("timeInUTC", utcTime);
+
+            return {
+              title: `Booked - ${utcTime}`,
+              start: combinedDateTime.toISOString(),
+              allDay: true,
+              color: "green",
+            };
+          }),
+      );
+
+    setEvents(filteredEvents);
+  };
+
+  const fetchSlots = async (start: string) => {
+    try {
+      const slotsData = await checkSlotAvailability(start);
+      setOriginalSlots(slotsData);
+      formatEvents(slotsData, filter);
+      lastFetchedDate.current = start;
+    } catch (error) {
+      console.error("Error fetching slot availability:", error);
+    }
+  };
+  useEffect(() => {
+    const firstDayOfMonth = moment().startOf("month").format("MM-DD-YYYY");
+    fetchSlots(firstDayOfMonth);
+  }, []);
+
+  const handleDatesSet = (info: any) => {
+    const firstDateOfMonth = moment(info.view.currentStart).format(
+      "MM-DD-YYYY",
+    );
+
+    // Avoid redundant API calls if the current month is already fetched
+    if (lastFetchedDate.current === firstDateOfMonth) {
+      return;
+    }
+
+    fetchSlots(firstDateOfMonth); // Call the API with the new month
+  };
+
+  const handleFilterChange = (newFilter: string) => {
+    setFilter(newFilter);
+    setHasSelectedTime(true);
+    if (originalSlots) {
+      formatEvents(originalSlots, newFilter);
+    }
+  };
+
+  const handleDateClick = (data: any) => {
+    const isDisabled = data.dayEl.classList.contains("disabled-cell");
+    if (isDisabled) {
+      return;
+    }
+
+    const clickedDate = moment(data.date).format("YYYY-MM-DD");
+
+    // Find the slot for the clicked date in originalSlots
+    const slotForDate: any = originalSlots
+      ? Object.values(originalSlots)?.find((slot: any) =>
+          slot?.date.startsWith(clickedDate),
+        )
+      : null;
+
+    if (!slotForDate || !slotForDate?.slots) {
+      return; // Exit if no slots are available
+    }
+
+    // Filter the slots based on the AM/PM filter
+    const filteredSlots = slotForDate?.slots?.filter((timeSlot: any) => {
+      const hour = parseInt(timeSlot.time.split(":")[0], 10);
+      return filter === "AM" ? hour < 12 : hour >= 12;
+    });
+
+    // Find the first available time slot
+    const firstAvailableSlot = filteredSlots?.find(
+      (timeSlot: any) => timeSlot.available,
+    );
+
+    if (!firstAvailableSlot) {
+      return; // Exit if no slots are available in the selected range
+    }
+
+    const newEventTime = firstAvailableSlot.time; // Use the time from the available slot
+    const newEventStart = `${clickedDate}T${newEventTime}:00`;
+
+    // Remove the previously added event
+    const updatedEvents = events.filter((event) => event.userAdded !== true);
+
+    // Add the new user event
+    const newEvent = {
+      title: `Engaged - ${moment(newEventStart).format("h:mm A")}`, // Add time in 12-hour format
+      start: newEventStart,
+      allDay: true,
+      color: "#74604B",
+      userAdded: true,
+    };
+
+    setEvents([...updatedEvents, newEvent]);
+    setIsButtonDisabled(false);
+    setSelectedDateTime(newEvent);
+  };
+
+  const dayCellClassNames = (date: any) => {
+    // Check if the current date is in originalSlots and slots is false
+    const dateStr = moment(date.date).format("YYYY-MM-DD"); // Format the date to match your keys
+    const slot: any = originalSlots
+      ? Object.values(originalSlots)?.find((slot: any) =>
+          slot?.date?.startsWith(dateStr),
+        )
+      : null;
+
+    const today = moment().startOf("day");
+
+    if (
+      moment(date.date).isBefore(today, "day") ||
+      (slot && slot?.slots === false) ||
+      slot?.slots === null
+    ) {
+      return ["disabled-cell"];
+    }
+    return [];
+  };
+
+  const handleProceed = () => {
+    if (selectedDateTime) {
+      const startTime = selectedDateTime?.start;
+      const date = new Date(startTime);
+      const formattedDateTime = new Date(
+        date.getTime() - date.getTimezoneOffset() * 60000,
+      ).toISOString();
+
+      router.push(`/booking?selectedDate=${formattedDateTime}`);
+    }
+  };
+
+  return (
+    <>
+      <div
+        className={cn(
+          "container mb-8",
+          !hasSelectedTime &&
+            "flex min-h-[100vh] flex-col items-center justify-center",
+        )}
+      >
+        {!hasSelectedTime ? (
+          <div className="text-center text-white">
+            <h3 className="mb-6 text-3xl">What time would you like to book?</h3>
+            <div className="flex justify-center gap-4">
+              <Button onClick={() => handleFilterChange("AM")}>Morning</Button>
+              <Button onClick={() => handleFilterChange("PM")}>Evening</Button>
+            </div>
+          </div>
+        ) : (
+          <>
+            <h2 className="text-center font-cormorant text-[4rem] text-white">
+              Calendar
+            </h2>
+
+            <div className="relative my-8">
+              {originalSlots && (
+                <div className="absolute left-4 top-3.5 flex items-center">
+                  <button
+                    className={cn(
+                      " ",
+                      filter === "AM" ? "text-black" : "text-black/40",
+                    )}
+                    onClick={() => handleFilterChange("AM")}
+                  >
+                    AM
+                  </button>
+
+                  <div className="mx-2 h-5 w-0.5 bg-black" />
+                  <button
+                    className={cn(
+                      " ",
+                      filter === "PM" ? "text-black" : "text-black/40",
+                    )}
+                    onClick={() => handleFilterChange("PM")}
+                  >
+                    PM
+                  </button>
+                </div>
+              )}
+
+              <FullCalendar
+                plugins={[dayGridPlugin, interactionPlugin]}
+                initialView="dayGridMonth"
+                events={events}
+                headerToolbar={{
+                  left: "",
+                  center: "prev,title,next",
+                  right: "",
+                }}
+                validRange={{
+                  start: new Date(
+                    startMonth.getFullYear(),
+                    startMonth.getMonth(),
+                    1,
+                  ),
+                }}
+                datesSet={handleDatesSet} // Triggered on initial render and navigation
+                dateClick={handleDateClick}
+                dayCellClassNames={dayCellClassNames} // Add custom classes to cells
+              />
+            </div>
+          </>
+        )}
+
+        {originalSlots && hasSelectedTime && (
+          <div className="mx-auto w-fit">
+            <Button disabled={isButtonDisabled} onClick={handleProceed}>
+              Proceed
+            </Button>
+          </div>
+        )}
+      </div>
+    </>
+  );
+};
+
+export default CalendarPage;
